requirements.txt
poe-api-wrapper==1.4.1
poe-api-wrapper[async]==1.4.1
redis==5.1.0b4
python-dotenv==1.0.1
nest_asyncio==1.6.0
langchain==0.1.20